+++ src/vstr/tests.rs
@@ -4,6 +4,7 @@ mod classify;
 mod emoji;
 mod encoding;
 mod humanize;
+mod inflection;
 mod matcher;
 mod path;
 #[cfg(feature = "pattern-regex")]